BACKPORT: bpf: inet_diag: Dump socket local storage
[ Upstream commits
5682d393b40e1fe7426a7b8c3471f05262f42010
0df6d32842b9a5f97a29ea90c8adc5cfac38341d
085c20cacf2b72991ce1c9d99a5e2f1d9e73bb68 ]

Accept INET_DIAG_REQ_SK_BPF_STORAGES in socket dump requests and
return the selected BPF socket-local-storage values in nested reply
attributes. An empty map list requests every storage value on a socket.

Parse diagnostic request attributes once at dump start and retain the
result in the netlink callback. Grow an empty reply buffer when a single
socket and its storage values do not fit, with the allocation bounded by
the maximum skb-backed kmalloc size.

[Mathias Gluszczynski: pass the dump callback alongside the older
inet_sk_diag_fill argument set instead of importing the broader socket
diagnostic API rewrite. Preserve the Linux 5.15 reply attribute number
by reserving its preceding ULP-info slot.]

+++ include/linux/netlink.h
@@ -129,7 +129,7 @@ struct netlink_callback {
 	/* the module that dump function belong to */
 	struct module		*module;
 	u16			family;
-	u16			min_dump_alloc;
+	u32			min_dump_alloc;
 	unsigned int		prev_seq, seq;
 	long			args[6];
 };
@@ -149,7 +149,7 @@ struct netlink_dump_control {
 	int (*done)(struct netlink_callback *);
 	void *data;
 	struct module *module;
-	u16 min_dump_alloc;
+	u32 min_dump_alloc;
 };
 
 extern int __netlink_dump_start(struct sock *ssk, struct sk_buff *skb,
